backend/requirements.txt
pymongo==4.3.3
bcrypt==4.0.1
PyJWT==2.6.0
numpy==1.24.3
opencv-python==4.7.0.72
PyWavelets==1.4.1
scipy==1.10.1
cryptography==40.0.2
Pillow==9.5.0
wave==0.0.2